Make sure VMs have SSH host keys
When using vagrant-libvirt then 'vagrant package' command runs
'virt-sysprep' which, among the other things, removes the SSH
host keys. Unfortunately, Ubuntu's sshd is not clever enough to
generate new host keys when none are available and silenty just
starts and pretends everything is okay. Except that it closes all
connections right away. So we need to make sure SSH host keys are
generated on boot if they are missing.