Added note about bad ARP entries
When trying to connect to buildslave from dev machine, I experienced
issues with bad entries in the ARP cache. Added a note on how I fixed
it.

Not sure if it was due to an ARP cache bug or a lingering ghost VM.
Hence, I added the note to the end of the README.md to not make two much
noice.

Co-authored-by: Craig Comstock <craig.comstock@northern.tech>
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>